% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_AUC.plot.R
\name{AUC.plot}
\alias{AUC.plot}
\title{to do}
\usage{
AUC.plot(auc, gSetName = "AUC distribution for this geneset",
  aucThr = max(auc), returnInfo = FALSE)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\examples{
#to do
}

