% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_buildRankings.R
\name{AUCell.buildRankings}
\alias{AUCell.buildRankings}
\title{to do}
\usage{
AUCell.buildRankings(exprMat, plotStats = TRUE, nCores = 1,
  verbose = TRUE)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\examples{
 # to do
}

