% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_calcAUC.R
\name{AUCell.calcAUC}
\alias{AUCell.calcAUC}
\title{to do}
\usage{
AUCell.calcAUC(geneSets, rankings, nCores = 1, aucMaxRank = 0.03 *
  nrow(rankings), verbose = TRUE)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\examples{
 #to do
}

