% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_AUCell.exploreThresholds.R
\name{AUCell.exploreThresholds}
\alias{AUCell.exploreThresholds}
\title{to do}
\usage{
AUCell.exploreThresholds(aucMatrix, seed = 987, thrP = 0.01, nCores = 1,
  plotHist = TRUE, densAdjust = 2, assignCells = FALSE, verbose = TRUE)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\details{
# required?
}
\examples{
 #to do
}

