% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_createViewerApp.R
\name{AUCell_createViewerApp}
\alias{AUCell_createViewerApp}
\title{Create AUCell viewer app}
\usage{
AUCell_createViewerApp(auc, thresholds = NULL, tSNE = NULL)
}
\arguments{
\item{auc}{AUC object returned by \code{\link{AUCell_calcAUC}}}

\item{thresholds}{Thresholds corresponding to each gene set (optional)}

\item{tSNE}{t-SNE coordinates for the cells (optional). 
The row names should correspond to the cell ID. 
The column names should be "tsne1" and "tsne2".}
}
\value{
Thresholds and cells selected within the app (as list).
}
\description{
Creates a Shiny app to explore AUCell results
}
\examples{
\dontrun{
# Create the Shiny app: 
aucellApp <- AUCell_createViewerApp(auc=cells_AUC, 
               thresholds=selectedThresholds,
               tSNE=cellsTsne)
               
# This object contains the $ui and $server required to lauch the app, e.g.:
savedSelections <- shinyApp(ui=aucellApp$ui, server=aucellApp$server)

# (How to launch the app might depend on the local settings)
options(shiny.host="0.0.0.0") 
savedSelections <- runApp(aucellApp) 
}
}
