% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_plotTSNE.R
\name{AUCell_plotTSNE}
\alias{AUCell_plotTSNE}
\title{Plot}
\usage{
AUCell_plotTSNE(tSNE, exprMat, cellsAUC = NULL, thresholds = NULL,
  cex = 1, alphaOn = 1, alphaOff = 0.2,
  borderColor = adjustcolor("lightgray", alpha = 0.1),
  offColor = "lightgray", plots = c("histogram", "binaryAUC", "AUC",
  "expression"), asPNG = FALSE, ...)
}
\arguments{
\item{tSNE}{t-SNE coordinates (e.g. \code{tSNE$Y})}

\item{exprMat}{Expression matrix}

\item{cellsAUC}{AUC (as returned by calcAUC)}

\item{thresholds}{Thresholds returned by AUCell}

\item{cex}{Scaling factor for the dots in the scatterplot}

\item{alphaOn}{Transparency for the dots representing "active" cells}

\item{alphaOff}{Transparency for the dots representing "inactive" cells}

\item{borderColor}{Border color for the dots (scatterplot)}

\item{offColor}{Color for the dots representing "inactive" cells}

\item{plots}{Which plots to generate? Select one or multiple: \code{plots=c("histogram", "binaryAUC", "AUC", "expression")}}

\item{asPNG}{Output each individual plot in a .png file?}

\item{...}{Other arguments to pass to \code{\link{hist}} function.}
}
\value{
Returns invisible: \code{cells_trhAssignment}
}
\description{
Plots the AUC histogram and t-SNE coloured by AUC, binary activity and TF expression
}
\details{
To avoid calculating thresholds, set thresholds to FALSE
}
\seealso{
List of vignettes included in the package: \code{vignette(package="SCENIC")}
}
