% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_aucellResults.R
\docType{class}
\name{aucellResults-class}
\alias{aucellResults-class}
\alias{aucellResults}
\alias{getAUC}
\alias{getRanking}
\alias{show}
\alias{show,aucellResults-method}
\alias{getAUC}
\alias{getAUC,aucellResults-method}
\alias{getRanking}
\alias{getRanking,aucellResults-method}
\title{Wrapper to the matrix that stores the AUC or the cell rankings.}
\usage{
\S4method{show}{aucellResults}(object)

getAUC(object)

getRanking(object)
}
\description{
This class is a wrapper for the AUC matrix and cell rankings.

The results are stored in the assays slot, but they can be accessed through the regular methods (i.e. nrow, rownames... )

Types:
 - "AUC": The assays contains the AUC for the gene-sets (or region-sets) & cells.
 - "ranking": The assays contains the gene rankings for each cell.

Methods: See examples.
}
