% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_getSetNames.R
\name{getSetNames}
\alias{getSetNames}
\title{Get gene set name (excluding number of genes/regions after space)}
\usage{
getSetNames(aucMat, patterns, startChr = "^", endChr = " |_")
}
\arguments{
\item{aucMat}{AUC matrix}

\item{patterns}{patterns}

\item{startChr}{Character to indicate the start (typically "^")}

\item{endChr}{Character at the end of the gene set name, i.e. space or "_" after a transcription factor name}
}
\value{
Returns the gene set name (i.e. selects the given pattern)
}
\description{
Returns the gene set name (i.e. selects the given pattern)
}
