% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_matrixWrapper.R
\docType{class}
\name{matrixWrapper-class}
\alias{getAuc}
\alias{getRanking}
\alias{matrixWrapper}
\alias{matrixWrapper-class}
\title{Wrapper to the matrix that stores the AUC or the cell rankings.}
\description{
This class is a wrapper for the AUC matrix and cell rankings.

The matrices are stored in the slot @matrix slot with the matrix, but they can be accessed through the regular methods (i.e. nrow, rownames... )

Types:
 - "AUC": The matrix contains the AUC for the gene-sets (or region-sets) & cells.
 - "ranking": The @matrix contains the gene rankings for each cell.

Methods: See examples.
}
\examples{
#####
# Fake AUC, normally this object is returned by calcAUC()
aucMatrix <- matrix(sample(1/1:30, 120, replace=TRUE), nrow=8)
 rownames(aucMatrix) <- paste("geneSet",1:nrow(aucMatrix), sep="")
 colnames(aucMatrix) <- paste("cell",1:ncol(aucMatrix), sep="")
fakeAUC <- matrixWrapper(matrix=aucMatrix, rowType="gene-set", colType="cell", matrixType="AUC")
#####

fakeAUC

# To subset & access the AUC or ranking slot:
fakeAUC[1:2,]
fakeAUC[,3:4]
getAuc(fakeAUC)[,1:5] # fakeAUC@matrix[,1:5]

# To subset the AUC or ranking, but keeping the class:
subset(fakeAUC, 1:2) # Default subsets on rows
subset(fakeAUC, 3:4, "cell")

# These methods are also available for this class:
dim(fakeAUC)
nrow(fakeAUC)
ncol(fakeAUC)
colnames(fakeAUC)
rownames(fakeAUC)

}

