% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_orderAUC.R
\name{orderAUC}
\alias{orderAUC}
\title{orderAUC}
\usage{
orderAUC(auc)
}
\arguments{
\item{auc}{AUC (as returned by calcAUC)}
}
\value{
gene-set names in the suggested order
}
\description{
Reorder the gene-sets based on AUC similarity
}
\examples{
# cellsAUC <- cellsAUC[orderAUC(cellsAUC),]
}
