% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_plotEmb_rgb.R
\name{plotEmb_rgb}
\alias{plotEmb_rgb}
\title{plotEmb_rgb}
\usage{
plotEmb_rgb(
  aucMat,
  embedding,
  geneSetsByCol,
  aucType = "AUC",
  aucMaxContrast = 0.8,
  offColor = "#c0c0c030",
  showPlot = TRUE,
  showLegend = TRUE,
  ...
)
}
\arguments{
\item{aucMat}{AUC matrix (continuous or binary)}

\item{embedding}{AUC matrix (continuous or binary)}

\item{geneSetsByCol}{Gene sets to plot}

\item{aucType}{"AUC" or "Binary"}

\item{aucMaxContrast}{To increase the AUC contrast decrease the value.}

\item{offColor}{Color por the cells completelly off. To deactivate (color as black), set to NULL.}

\item{showPlot}{Whether to plot the coloured embeddings.}

\item{showLegend}{Whether to plot add a legend to the plot.}

\item{...}{Other arguments to pass to the \code{plot} function.}
}
\value{
The cell colors (invisible)
}
\description{
Colors the embeddings (t-SNE/Umap) based on the activity of 3 (groups of) geneSets
}
