% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_updateAucellResults.R
\name{updateAucellResults}
\alias{updateAucellResults}
\title{Update AUCell results}
\usage{
updateAucellResults(oldAucObject, objectType = "AUC")
}
\arguments{
\item{oldAucObject}{Object to update}

\item{objectType}{Either "AUC" or "ranking" indicating the object type}
}
\value{
Updated version of the object as \code{\link{aucellResults}}.
}
\description{
Updates the AUC scores provided by AUCell from a previous version.
}
\examples{
oldAuc <- matrix(data=1:2000, nrow=50, ncol=40)
updateAucellResults(oldAuc)
}
