% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_buildRankings.R
\name{AUCell.buildRankings}
\alias{AUCell.buildRankings}
\title{Build gene expression rankings for each cell}
\usage{
AUCell.buildRankings(exprMat, plotStats = TRUE, nCores = 1,
  verbose = TRUE)
}
\arguments{
\item{exprMat}{Expression matrix (genes as rows, cells as columns)}

\item{plotStats}{Should the function plot the expression boxplots/histograms? (TRUE / FALSE). These plots can also be produced with the function \code{\link{plotGeneCount}}.}

\item{nCores}{Number of cores to use for computation.}

\item{verbose}{Should the function show progress messages? (TRUE / FALSE)}
}
\value{
data.table of genes (row) by cells (columns) with the ranking of the gene within the cell.
}
\description{
Builds the "rankings" for each cell: expression-based ranking for all the genes in each cell.

The genes with same expression value are shuffled. Therefore, genes with expression '0' are randomly sorted at the end of the ranking.

These "rankings" can be seen as a new representation of the original dataset. Once they are calculated, they can be saved for future analyses.
}
\details{
It is important to check that most cells have at least the number of expressed/detected genes that are going to be used to calculate the AUC (`aucMaxRank` in `calcAUC()`). The histogram provided by `AUCell.buildRankings()` allows to quickly check this distribution. `plotGeneCount(exprMatrix)` allows to obtain only the plot before building the rankings.
}
\examples{
# This example is run using a fake expression matrix.
# Therefore, the output will be meaningless.

############# Fake expression matrix #############
set.seed(123)
exprMatrix <- matrix(data=sample(c(rep(0, 5000), sample(1:3, 5000, replace=TRUE))),
                     nrow=20)
rownames(exprMatrix) <- paste("Gene", 1:20, sep="")
colnames(exprMatrix) <- paste("Cell", 1:500, sep="")
dim(exprMatrix)
##################################################

cells_rankings <- AUCell.buildRankings(exprMatrix, plotStats=TRUE)
cells_rankings
}
\seealso{
Next step in the workflow: \code{\link{AUCell.calcAUC}}.

See the package vignette for examples and more details: \code{vignette("AUCell")}
}

