% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_cellsAUC.R
\docType{class}
\name{cellsAUC-class}
\alias{cellsAUC-class}
\alias{cellsAUC}
\title{Class to store AUC}
\description{
Class with only one slot: @AUC, a matrix which contains the AUC for the gene-sets & cells.
}
